Move tests to different file
